requirements.txt
flask
pillow
pytesseract
pandas
fuzzywuzzy
python-Levenshtein
matplotlib
wordcloud
